Move HTTP functions to pkg/handlerutil
Moving forward, they will be needed to implement Helm 3 support (see
#1309), and they were in an internal package, so they couldn't be
imported outside the `cmd/tiller-proxy` directory. Because they were all
pure (or, in the case of ServeHTTP, at least very simple), I felt
confident moving them to another file and reusing them.
